Adept Chapter: Simulation update
Spotted an error with the way that follow-up time and duration is calculated. So re ran all the simulations with a TITE component (PO-TITE-CRM, alternative TITE-CRM and the modified PO-TITE-CRM designs).  Updated the tables in the text accordingly. Also, updated the text for the standard PO-TITE-CRM simulations to reflec the new results. Now just need to update the text for the modified /alternative designs (section 2.4).