Renamed chapter responded to comments started a plan
Since no longer using protect trial I renamed the folder to WT for Wages and Tait.

Addressed comments Kristian highlighted. It involved changing the code that generates a table to calculate a percentage instead of a number.

Also, Included the code I used to generate some basic examples of the Wages and Tait design with randomisation.

Started a plan for the draft of the chapter.