WT Evaluation of Design
Started the section of the chapter looking at evaluating and exploring the design.

Included introduction to the section explaining the design being used.

Also wrote section introducing the exploration sims and what scenarios are being run.

Wrote code for exploration and comparison sims. Need to include scenarios 9-12 for the comparison code.

Simulations are currently running. I'll need to run scenarios 9-12 for each combo for the exploration then save that data.

Made some slight edits to the introduction explained the design of another methodology paper.