WT Evaluation of the Design + Discussion
Wrote up results from the alternative design.

Ran additional simulations looking at the case when control is active. Looked at two different designs with different efficacy skeletons. Wrote the introduction for this section and produced the tables. Just need to write up the results.

Started writing the discussion, only a page so far need to fill it out with the comparisons of alternative designs and the active control segments.